Ajout des permissions dans l'instance User Ajout de la permission dans fetch_data
